clippy fix `unnessecary cast`

+++ lib/legacy/zstd_v05.rs
@@ -198,7 +198,7 @@ unsafe fn ZSTDv05_wildcopy(
 ) {
     let mut ip = src as *const u8;
     let mut op = dst as *mut u8;
-    let oend = op.offset(length as isize);
+    let oend = op.offset(length);
     loop {
         ZSTDv05_copy8(op as *mut core::ffi::c_void, ip as *const core::ffi::c_void);
         op = op.offset(8);
